clean: Simplify atomic compiler passes. Update docs
