+++ .gitignore
@@ -180,3 +180,4 @@ tests/test.bat
 
 .pre-commit-config.yaml
 *.antlr/
+examples/*
