+++ src/oqd_core/compiler/atomic/cfg_passes/__init__.py
@@ -1,9 +1,8 @@
-from .walk import iter_stmt_blocks
-from .scalar_env import canonicalize_scalar_expr, canonicalize_scalars_cfg, resolve_scalar_expr
+from .walk import iter_stmt_blocks, canonicalize_declarations_cfg
+from .resolve import resolve_scalar_expr
 
 __all__ = [
     "iter_stmt_blocks",
-    "canonicalize_scalar_expr",
-    "canonicalize_scalars_cfg",
+    "canonicalize_declarations_cfg",
     "resolve_scalar_expr",
 ]
