+++ src/oqd_core/compiler/atomic/verify/__init__.py
@@ -1,5 +1,4 @@
-from .passes import verify_constant_pulse_durations, verify_pulse_target_dim
+from .passes import verify_pulse_target_dim
 __all__ = [
-    "verify_constant_pulse_durations",
     "verify_pulse_target_dim",
 ]
